I have changed the program to not have any global constants.
Instead, the constants are defined once in a structure called
program_const and then passed around as a pointer to all functions that
need it. This will make it easier to change the constants (also from
python) at runtime. To do this all constants::XX expressions in the code
have been changes and many of the argument lists. Since the constants
now are not const the function LS_Solver::get_mu_q_on_shell is no longer
static and hase been changed to be called using LS_Solver_->... instead.