I have made the arguments in potential_mwpc.cpp consistent with the convention that p_out,p_in order.
I have run all the tests of the code and everything is exactly the same
as before the change.

I have also inspected HO matrix elements generated in LO WPC and see no
difference compared to before the changes.

With these changes the logic of  the code is better, but all potential
outputs should be unaltered.